A clean property is now added each time a cell is created.
This allows for safe updating of property values without overwriting
values of global properties.  In the case of read-only properties,
these properties cannot be updated or their value changed and only a
single instance of each read-only property is stored.

Default properties are now tagged by the property and result in an
error if the user tries to remove them from the PropertyList.