Updated Site::setType to check the allowed site types and throw an error if the types don't match.
In lieu of the potential slow down, I added a new setTypeUnchecked method
which will forgo the check.  I was going to update the setType calls in
the tcp reader, but a TODO task already existed for checking the type
to ensure it is valid, so I left the check in.

I don't foresee a big slow down since this is not a commonly used method
in most situations the size of the list are normally no more than 2-3
elements large.