Added classes for creating compressed XDLRC files.
Unlike device files, these files have all information necessary to
(almost) fully restore the XDLRC file.  The compressed XDLRC files are
on the order of 50-100MB even for the larger V7 devices.  This should
allow for storing and sharing XDLRC files.

A means of reading in the compressed file with the XDLRCParserListener
classes to duplicate the XDLRCParser functionality.  This should allow
the compressed files to be usable with the DeviceGenerator and other
utilities requiring the XDLRC files.  They should also parse significantly
faster than the XDLRC files.

Also included in this commit are updates to the HashPool class and
an XDLRCRegurgitatorListener.  The regurgitator is a parser listener that
simply spits outside an XDLRC file generated from the passed in parse
tokens.  Its utility is mainly for testing.

The HashPool has been updated
with two new methods.  A constructor taking an initial size to reduce
growing the internal structure sizes (in use, the HashPools can grow
pretty large).  The add2 method acts identically to add, but returns
the ordinal value of the passed in object instead of the pooled object.
The getEnumeration method has also been updated to return an Integer
object instead of an int primitive to help reduce the number of Integer
objects laying around from reboxing the primitives returned by the method.